Use serialized output for tests
We originally used var_export() which gave us human-friendly expected output.
However, it turns out that var_export() outputs some floating point values
differently in PHP 7+ which caused those tests to fail.

print_r() was not an option because it does not output the correspoding types;
a float of 1.0 and an integer of 1 would appear identical in the output.

var_dump() doesn't have the ability to return the dump. Output buffers are a
possible workaround but they were painfully slow.

serialize() was therefore chosen because it's somewhat readable and also
perfectly conveys the types and values.